requirements.txt
aiogram==2.18
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
contextlib2==21.6.0
discord-webhook==0.14.0
discord.py==1.7.3
frozenlist==1.2.0
idna==3.3
jsonschema==4.4.0
multidict==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
tomli==1.2.3
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
